+++ packages/core/src/service/helpers/entity-hydrator/entity-hydrator.service.ts
@@ -122,11 +122,24 @@ export class EntityHydrator {
                     .filter(item => this.isTranslatable(item.entity))
                     .map(item => item.relation.split('.'));
 
-                Object.assign(
+                this.assignSettableProperties(
                     target,
                     translateDeep(target as any, ctx.languageCode, translateDeepRelations as any),
                 );
             }
+        }
+        return target;
+    }
+
+    private assignSettableProperties<Entity extends VendureEntity>(target: Entity, source: Entity) {
+        for (const [key, descriptor] of Object.entries(Object.getOwnPropertyDescriptors(target))) {
+            if (typeof descriptor.get === 'function' && typeof descriptor.set !== 'function') {
+                // If the entity property has a getter only, we will skip it otherwise
+                // we will get an error of the form:
+                // `Cannot set property <name> of #<Entity> which has only a getter`
+                continue;
+            }
+            target[key as keyof Entity] = source[key as keyof Entity];
         }
         return target;
     }
@@ -143,11 +156,11 @@ export class EntityHydrator {
         for (const relation of options.relations.slice().sort()) {
             if (typeof relation === 'string') {
                 const parts = relation.split('.');
-                let entity: any = target;
+                let entity: Record<string, any> | undefined = target;
                 const path = [];
                 for (const part of parts) {
                     path.push(part);
-                    if (entity[part]) {
+                    if (entity && entity[part]) {
                         entity = Array.isArray(entity[part]) ? entity[part][0] : entity[part];
                     } else {
                         const allParts = path.reduce((result, p, i) => {
@@ -158,6 +171,7 @@ export class EntityHydrator {
                             }
                         }, [] as string[]);
                         missingRelations.push(...allParts);
+                        entity = undefined;
                     }
                 }
             }